pyproject.toml
[project]
name = "x-vits"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "joblib>=1.4.2",
    "lightning>=2.4.0",
    "loguru>=0.7.2",
    "matplotlib>=3.9.2",
    "numba>=0.60.0",
    "pandas>=2.2.3",
    "pyworld-prebuilt>=0.3.4.1",
    "scipy>=1.14.1",
    "soundfile>=0.12.1",
    "tensorboard>=2.18.0",
    "torchaudio>=2.4.1",
    "wandb>=0.18.3",
    "numpy<2",
    "jaxtyping>=0.2.34",
    "rich>=13.9.2",
    "rotary-embedding-torch>=0.8.4",
    "transformers>=4.45.2",
    "sentencepiece>=0.2.0",
    "einops-exts>=0.0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.metadata]
allow-direct-references = true


[tool.ruff]
line-length = 100
